Add files via upload
The Human Resource Management is concerned about the number of employees leaving. They assume that employees who receive training at the time of probation will tend to be more loyal to the company when compared to those who do not receive training training. You as an HRM analyst are appointed to test this assumption.